compose.tsx gives back 700 lines
The composer sits under a frozen line cap, and the slices that add the
communication status line, the review drawer and the send command all
have to land inside it. A file one line under its waiver has no room for
any of them.

Three self-contained regions move out, unchanged:

composeschedule.tsx is choosing WHEN a message goes — three presets, a
calendar, and the wall-clock arithmetic that keeps a preset meaning what
it says in the reader's own zone. It touches no draft, no recipient and
no send.

composerelink.tsx is moving a filed message to the record it belongs to.
RELINK_KINDS lives there because it is the vocabulary BOTH surfaces
speak: the drawer files under one of these kinds and the timeline moves
one between them, and a second spelling is how the two stop agreeing.

composedraftcontext.tsx is the account-started draft: the pickers that
say what a message is about, the panel that says what the model read, and
the offer that turns a refusal into a next step. Three sides of one
question — what does this draft stand on.

MOVED, NOT REWRITTEN. Every line that lands is a line that was there, and
no test changed: 8755 pass exactly as before, which is the only proof a
refactor of this shape can offer. RELINK_KINDS, RelinkKind and RelinkModal
are re-exported from compose.tsx so the ten surfaces importing them keep
one import path — an extraction is not an invitation to touch ten call
sites.

3717 lines to 3014, and the waiver is ratcheted down to match so the
space cannot quietly be spent on something else.
